requirements.txt
flask==2.3.3
pyspark==3.5.1
matplotlib==3.8.0
pandas==2.2.2
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
pyotp==2.9.0
twilio==8.10.0
flask-wtf==1.2.1
wtforms==3.1.2
pyodbc==5.2.0
sqlalchemy==2.0.31
python-dotenv==1.1.1
